Move memory write to be a shared module
