+++ src/linux/dso_debug.rs
@@ -1,10 +1,6 @@
 use crate::{
-    linux::{
-        auxv_reader::AuxvType,
-        errors::SectionDsoDebugError,
-        ptrace_dumper::PtraceDumper,
-        sections::{write_string_to_location, Buffer, MemoryArrayWriter, MemoryWriter},
-    },
+    linux::{auxv_reader::AuxvType, errors::SectionDsoDebugError, ptrace_dumper::PtraceDumper},
+    mem_writer::{write_string_to_location, Buffer, MemoryArrayWriter, MemoryWriter},
     minidump_format::*,
 };
 use std::collections::HashMap;
