+++ src/linux/errors.rs
@@ -1,4 +1,5 @@
 use crate::maps_reader::MappingInfo;
+use crate::mem_writer::MemoryWriterError;
 use crate::thread_info::Pid;
 use goblin;
 use thiserror::Error;
@@ -118,16 +119,6 @@ pub enum DumperError {
     TryFromIntError(#[from] std::num::TryFromIntError),
     #[error("Maps reader error")]
     MapsReaderError(#[from] MapsReaderError),
-}
-
-#[derive(Debug, Error)]
-pub enum MemoryWriterError {
-    #[error("IO error when writing to DumpBuf")]
-    IOError(#[from] std::io::Error),
-    #[error("Failed integer conversion")]
-    TryFromIntError(#[from] std::num::TryFromIntError),
-    #[error("Failed to write to buffer")]
-    Scroll(#[from] scroll::Error),
 }
 
 #[derive(Debug, Error)]
